Add Sha2LongDigest abstraction for SHA-384 and SHA-512
