+++ library/sha2/sha2-384/api/sha2-384.api
@@ -1,4 +1,4 @@
-public final class org/kotlincrypto/hash/Sha384 : org/kotlincrypto/core/Digest {
+public final class org/kotlincrypto/hash/Sha384 : org/kotlincrypto/hash/Sha2LongDigest {
 	public fun <init> ()V
 }
