+++ library/sha2/sha2-512/api/sha2-512.api
@@ -1,4 +1,4 @@
-public final class org/kotlincrypto/hash/Sha512 : org/kotlincrypto/core/Digest {
+public final class org/kotlincrypto/hash/Sha512 : org/kotlincrypto/hash/Sha2LongDigest {
 	public fun <init> ()V
 }
