+++ library/sha2/sha2-512/build.gradle.kts
@@ -23,7 +23,7 @@ kmpConfiguration {
         common {
             sourceSetMain {
                 dependencies {
-                    api(libs.kotlincrypto.core.digest)
+                    api(project(":library:sha2:sha2-digest-long"))
                 }
             }
             sourceSetTest {
