+++ library/sha2/sha2-digest-long/.gitignore
@@ -1,0 +1,1 @@
+build/
