+++ library/sha2/sha2-digest-long/gradle.properties
@@ -1,0 +1,16 @@
+# Copyright (c) 2023 Matthew Nelson
+#
+# Licensed under the Apache License, Version 2.0 (the "License");
+# you may not use this file except in compliance with the License.
+# You may obtain a copy of the License at
+#
+#     http://www.apache.org/licenses/LICENSE-2.0
+#
+# Unless required by applicable law or agreed to in writing, software
+# distributed under the License is distributed on an "AS IS" BASIS,
+# WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+# See the License for the specific language governing permissions and
+# limitations under the License.
+POM_ARTIFACT_ID=sha2-digest-long
+POM_NAME=KotlinCrypto Sha2 Long Digest Abstraction
+POM_DESCRIPTION=Core component for SHA-384 and SHA-512 Digests
